gestion des entiers, fusion des ajouts de compte, correction des choix admin en conséquenc, verification avant suppression

+++ src/fonctions principales/Suppression_adherent.c
@@ -4,7 +4,7 @@
 #include <mysql/mysql.h>
 #include <unistd.h>
 #include "qui_et_quand.h"
-#include "suppression_adherent.h"
+#include "prototype_admin_general.h"
 
 void Suppression_adherent(const char *login) {
     char *username, *time_str;
@@ -19,6 +19,24 @@ void Suppression_adherent(const char *login) {
     // Écrivez l'en-tête du log
     fprintf(log_file, "Utilisateur: %s, Date et heure: %s\n", username, time_str);
 
+    // Vérifiez si l'utilisateur existe
+    char query_select[500];
+    sprintf(query_select, "SELECT * FROM Utilisateur WHERE ID_Utilisateur = '%s'", login);
+    if (mysql_query(conn, query_select)) {
+        fprintf(log_file, "Erreur lors de la vérification de l'utilisateur dans la base de données: %s\n", mysql_error(conn));
+        fclose(log_file);
+        free(time_str);
+        return;
+    }
+
+    MYSQL_RES *result = mysql_store_result(conn);
+    if (mysql_num_rows(result) == 0) {
+        printf("L'utilisateur avec le login '%s' n'a pas été trouvé dans la base de données.\n", login);
+        fclose(log_file);
+        free(time_str);
+        return;
+    }
+    
     // Suppression de l'utilisateur de la base de données
     char query[500];
     sprintf(query, "DELETE FROM Utilisateur WHERE ID_Utilisateur = '%s'", login);
